+++ lib/ferrum/browser/options/base.rb
@@ -26,11 +26,11 @@ module Ferrum
         end
 
         def except(*keys)
-          to_h.reject { |n, _| keys.include?(n) }
+          to_h.except(*keys)
         end
 
         def detect_path
-          Binary.find(self.class::PLATFORM_PATH[Utils::Platform.name])
+          Binary.find(self.class::PLATFORM_PATH[Utils::Platform.platform_name])
         end
 
         def merge_required(flags, options, user_data_dir)